Move the packetblaster lwaftr code from src/apps/ to the command directory
